pyproject.toml
[project]
name = "ai-agent-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.31",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.3.8",
    "langsmith>=0.4.16",
    "mcp>=1.13.1",
    "python-dotenv>=1.1.1",
]
